recon: the engine compiles for x86-64; fix Clang tree-wide build
Two independent results.

1. Clang now builds the whole tree, not just the engine.

All five HLTV sub-projects passed -flto at compile but not at link. GCC
gets away with it because GNU ld auto-loads the LTO plugin for GCC's IR
objects; clang++ emits pure LLVM bitcode and the link fails with "file
format not recognized". Adding -flto to LINK_FLAGS is correct for both.
The README has claimed Clang support for a long time; it was only true for
the engine. All 8 artifacts now build under Clang, 33/33 tests pass, and
GCC is unaffected.

2. An actual -m64 compile, to derive the port's task list from compiler
output rather than source reading.

All 101 translation units build as ELF 64-bit x86-64 objects with ZERO
source-level compile errors. It took -m64, PIC back on, JIT and SSE off,
dropping the 32-bit-only aelf32 (strtools.h has a complete libc fallback),
and exactly one source change: info.cpp:98, where s - value is ptrdiff_t -
int on i386, long on LP64 - so Q_min stops deducing a common type.

This reframes the brief's claim that a 64-bit build is "a genuine port, not
a flag removal". At the compile stage it very nearly is a flag removal. The
work is at link and runtime instead, which is a different task list.

After dropping aelf32 the single remaining link blocker is libsteam_api.so,
which exists only as 32-bit. The stub surface is exactly 15 external
Steamworks symbols - small enough that an insecure/LAN development stub is
a few hours' work, which is the brief's own risk-register mitigation.

The compiler surfaced two real LP64 defects. sys_dll.cpp:697 declares
FindAddressInTable(extensiondll_t*, uint32 function) and casts that uint32
back to a pointer for dladdr - a code address round-tripped through 32 bits,
which the portability sweep missed. sv_main.cpp:3213 computes
2044 - sizeof(data) and underflows, but produces the identical -2052 on
i386, so it is pre-existing and architecture-neutral; deliberately not
changed.

REHLDS_TARGET_BITS=64 is labelled experimental and does not produce a
working server. Every blocker that matters - string_t truncation, edict
layout, the .hpk format - is runtime, not compile-time, and needs a 64-bit
ReGameDLL before it can even be reached.

32-bit GCC release and Unittests builds verified unchanged: 33/33 pass.

+++ rehlds/engine/info.cpp
@@ -95,7 +95,10 @@ const char* EXT_FUNC Info_ValueForKey(const char *s, const char *lookup)
 		while (*s != '\\' && *s != '\0')
 			s++;
 
-		size_t valueLen = Q_min(s - value, MAX_KV_LEN - 1);
+		// s - value is ptrdiff_t, which is int on i386 but long on LP64, so the two
+		// arguments stop having a common type there and Q_min (a template) fails to
+		// deduce. Both values are non-negative and small; make the type explicit.
+		size_t valueLen = Q_min((size_t)(s - value), (size_t)(MAX_KV_LEN - 1));
 
 		if (keyLen == lookupLen && !Q_strncmp(key, lookup, lookupLen))
 		{
